test(tabs): migrate oruga vue 2 test casse
